shop panel and shop image ok

+++ src/components/templates/panel/factors/AllInvoice.js
@@ -1,4 +1,3 @@
-// InvoiceManage.jsx
 "use client";
 import { useState, useEffect, useCallback } from "react";
 import FormTemplate from "@/templates/generalcomponnents/formTemplate";
@@ -7,21 +6,38 @@ import InvoiceCard from "./InvoiceCard";
 import EditSvg from "@/module/svgs/EditSvg";
 import DeleteSvg from "@/module/svgs/DeleteSvg";
 import { GetShopInvocesByShopId } from "./InvoicesActions";
-import { deleteInvoiceAction } from "./invoiceItemsServerActions2";
+import { deleteInvoiceAction } from "./invoiceItemsServerActions";
+import { GetShopLogos } from "@/templates/Shop/ShopServerActions";
 
 function InvoiceManage() {
   const [selectedInvioce, setSelectedInvoice] = useState(null);
   const params = useParams();
+  const [BGImage, setbGImage] = useState([]);
   const { ShopId } = params;
   const [invoices, setInvoices] = useState([]);
+  const getShopPanelImage = useCallback(async () => {
+      try {
+        if (!ShopId) {
+          console.error("نام یکتای فروشگاه موجود نیست.");
+          return;
+        }
+  
+        const response = await GetShopLogos(ShopId);
+  
+        setbGImage(response.logos.backgroundPanelUrl);
+      } catch (error) {
+        console.error("Error fetching banners:", error);
+      }
+    }, [ShopId]);
+  
+    useEffect(() => {
+      getShopPanelImage();
+    }, [ShopId]);
   
   const refreshInvioces = useCallback(async () => {
     try {
       const response = await GetShopInvocesByShopId(ShopId);
-      
       setInvoices(response.Invoices);
-      // فرض می‌کنیم هر فاکتور دارای یک فیلد contacts که آرایه‌ای از مخاطبان است
-     
     } catch (error) {
       console.error("Error fetching invoices:", error);
     }
@@ -31,56 +47,43 @@ function InvoiceManage() {
     refreshInvioces();
   }, [refreshInvioces]);
 
-
-
-
-
-/////////////////////////////////////////////
-
-
-
   const handleDeleteInvoice = useCallback(async (InvoiceID) => {
+    const isConfirmed = window.confirm("آیا از حذف این فاکتور اطمینان دارید؟");
+    if (!isConfirmed) return; // اگر کاربر لغو کند، ادامه نده
     let res = await deleteInvoiceAction(InvoiceID,ShopId);
-    
     if (res.success) {
       const updatedInvoices = invoices.filter((invoice) => invoice._id !== InvoiceID);
-      setInvoices(updatedInvoices);
-      
-     
+      setInvoices(updatedInvoices);    
     } else {
       console.error("Error deleting invoice:", res.status);
     }
   }, [invoices]);
-
-
-
   const handleEditClick = (invoice) => {
     setIsOpenAddInvoice(true);
     setSelectedInvoice(invoice);
   };
+  // const handleOverlayClick = (e) => {
+  //   if (e.target === e.currentTarget) {
+  //     setIsOpenAddInvoice(false);
+  //     setSelectedInvoice(null);
+  //     setIsOpenContactsList(false);
+  //     setSelectedContacts([]);
+  //     setUserListButtonName("");
+  //   }
+  // };
 
-  const handleOverlayClick = (e) => {
-    if (e.target === e.currentTarget) {
-      setIsOpenAddInvoice(false);
-      setSelectedInvoice(null);
-      setIsOpenContactsList(false);
-      setSelectedContacts([]);
-      setUserListButtonName("");
-    }
-  };
-
-  const handleCloseModal = () => {
-    setIsOpenAddInvoice(false);
-    setSelectedInvoice(null);
-  };
+  // const handleCloseModal = () => {
+  //   setIsOpenAddInvoice(false);
+  //   setSelectedInvoice(null);
+  // };
 
 
-  const handleSubmit = async (formData) => {
-    await refreshInvoices();
-  };
+  // const handleSubmit = async (formData) => {
+  //   await refreshInvoices();
+  // };
 
   return (
-    <FormTemplate>
+    <FormTemplate BGImage={BGImage}>
          <div className="bg-white bg-opacity-95 dark:bg-zinc-700 dark:bg-opacity-95 shadow-normal rounded-2xl mt-36">
         <div className="hidden">
           <EditSvg />
